+++ src/taxaplease/cli.py
@@ -13,13 +13,13 @@ def init_argparser():
     )
 
     parser.add_argument(
-        "--database_directory",
-        "-db_dir",
-        dest="database_directory",
+        "--database",
+        "-db",
+        dest="database",
         type=str,
         required=False,
         default=None,
-        help="set path to custom database folder",
+        help="Specify path to taxaPlease sqlite database file",
     )
 
     ## subparsers are mutually exclusive by default
@@ -183,7 +183,7 @@ def handle_taxonomy_request(args, taxapleaseObj):
 def main():
     args = init_argparser().parse_args()
 
-    tp = TaxaPlease(args.database_directory)
+    tp = TaxaPlease(database=args.database)
 
     result = None